+++ CHANGELOG.md
@@ -5,6 +5,8 @@
   + Deprecation messages are now supported for most elements.
   + Documentation references to `null` value are now replaced with `nil` in Swift.
   + Constants of struct types are now supported.
+- Breaking changes:
+  + Struct constructors that initialize "internal" fields are now also "internal" (Java, Swift).
 
 ## 4.7.1
 - Release date: 2019-07-01